+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2019-21-09
+### Added
+  - New setting - `platformSettings.autoLoad`.
+  - Inheritance support.
+
+### Modified
+  - Improved settings.
+
 ## [0.0.1] - 2019-21-09
 ### Added
-- Base functionality added.
-- New command - `extension.refreshSettings`.
+  - Base functionality added.
+  - New command - `extension.refreshSettings`.